adding sample drive program
